extract ToSingleGuardianSubscription
